+++ src/composables/providers/index.ts
@@ -1,5 +1,0 @@
-export * from './use-provider'
-export * from './use-metamask'
-export * from './use-coinbase'
-export * from './use-phantom'
-export * from './use-solflare'