+++ src/consts/chains.consts.ts
@@ -10,6 +10,16 @@ export const POLYGON_MAINNET_CHAIN: ChainInfo = {
   blockExplorerUrl: 'https://polygonscan.com/',
 }
 
+export const SEPOLIA_CHAIN: ChainInfo = {
+  rpcUrl: 'https://rpc2.sepolia.org',
+  blockExplorerUrl: 'https://sepolia.etherscan.io/',
+}
+
+export const GOERLI_CHAIN: ChainInfo = {
+  rpcUrl: 'https://ethereum-goerli.publicnode.com',
+  blockExplorerUrl: 'https://goerli.etherscan.io/',
+}
+
 export const Q_TESTNET_CHAIN: ChainInfo = {
   rpcUrl: 'https://rpc.qtestnet.org',
   blockExplorerUrl: 'https://explorer.qtestnet.org/',