+++ src/types/api/book.types.ts
@@ -27,3 +27,25 @@ export type Book = JsonApiRecordBase<'books'> & {
 export type CreateBookResponse = JsonApiRecordBase<'create_book'> & {
   book_id: string
 }
+
+export type FullBookInfo = Book & {
+  tokenContract: string
+  tokenName: string
+  tokenSymbol: string
+  pricePerOneToken: string
+  minNFTFloorPrice: string
+  voucherTokensAmount: string
+  voucherTokenContract: string
+  fundsRecipient: string
+  isNFTBuyable: boolean
+  isDisabled: boolean
+  isVoucherBuyable: boolean
+}
+
+export type BaseBookInfo = Book & {
+  pricePerOneToken: string
+  isDisabled: boolean
+  tokenContract: string
+  tokenName: string
+  tokenSymbol: string
+}