+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Please check our [developers guide](https://gitlab.com/tokend/developers-guide)
 for further information about branching and tagging conventions.
 
+## [1.5.1] - 2023-08-15
+#### Changed
+- Interaction with providers moved on [@distributedlab/w3p](https://github.com/distributed-lab/web-kit/tree/main/packages/w3p)
+- Book types refactored
+
 ## [1.5.0] - 2023-05-12
 #### Added
 - Deploying/editing on multiple chains at once
@@ -91,7 +96,8 @@ for further information about branching and tagging conventions.
 #### Under the hood changes
 - Merged from https://gitlab.com/distributed_lab/frontend/vue-template/-/commit/905e060b66b4fbdfa4bf102ac6c27e8edd11dbc2
 
-[Unreleased]: https://github.com/dl-nft-books/admin-panel/compare/v1.5.0...main
+[Unreleased]: https://github.com/dl-nft-books/admin-panel/compare/v1.5.1...main
+[1.5.1]: https://github.com/dl-nft-books/admin-panel/compare/v1.5.0...v1.5.1
 [1.5.0]: https://github.com/dl-nft-books/admin-panel/compare/v1.4.0-rc.1...v1.5.0
 [1.4.0-rc.1]: https://github.com/dl-nft-books/admin-panel/compare/v1.3.0...v1.4.0-rc.1
 [1.3.0]: https://github.com/dl-nft-books/admin-panel/compare/v1.2.0...v1.3.0